requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
